Use the new gog_guru_get_help_button to solve [#394181] until we split the
2007-07-18  Jody Goldberg <jody@gnome.org>

	* src/sheet-object-graph.c (sheet_object_graph_guru) : Use the new
	  gog_guru_get_help_button to solve [#394181] until we split the docs
	  between gnumeric and goffice.

svn path=/trunk/; revision=15688

+++ plugins/excel/xlsx-read.c
@@ -282,7 +282,7 @@ attr_int (GsfXMLIn *xin, xmlChar const **attrs,
 	  int *res)
 {
 	char *end;
-	int tmp;
+	long tmp;
 
 	g_return_val_if_fail (attrs != NULL, FALSE);
 	g_return_val_if_fail (attrs[0] != NULL, FALSE);
@@ -881,7 +881,7 @@ xlsx_get_num_fmt (GsfXMLIn *xin, char const *id)
 	XLSXReadState	*state = (XLSXReadState *)xin->user_state;
 	GOFormat *res = g_hash_table_lookup (state->num_fmts, id);
 	char *end;
-	int i;
+	long i;
 
 	if (NULL != res)
 		return res;
